backend/requirements.txt
fastapi
uvicorn
transparent-background
pillow
